+++ dash_app/callbacks/mafgraph.py
@@ -2,23 +2,20 @@ from dash_app.components import tools
 from ..server import app
 from dash.dependencies import Input, Output, State
 from ..layout.layout_ids import *
+from ..components import mafgraph
 
-
-@app.callback(
-    Output(id_mafgraph_graph, 'elements'),
-    [Input(id_mafgraph_hidden, 'children')],
-    [State(id_mafgraph_graph, 'elements')]
-)
-def update_poagraph(jsonified_mafgraph_elements: str, elements):
-    if not jsonified_mafgraph_elements:
+@app.callback(Output(id_mafgraph_graph, 'elements'),
+              [Input(id_pangenome_hidden, 'children')],
+              [State(id_mafgraph_graph, 'elements')])
+def show_input_vis(jsonified_pangenome, mafgraph_elements):
+    if not jsonified_pangenome:
         return []
-
-    nodes = tools.unjsonify_builtin_types(jsonified_mafgraph_elements[0]['props']['children'])
-    edges = tools.unjsonify_builtin_types(jsonified_mafgraph_elements[1]['props']['children'])
-    elements = []
-    elements.extend(nodes)
-    elements.extend(edges)
-    return elements
+    jsonpangenome = tools.unjsonify_jsonpangenome(jsonified_pangenome)
+    mafgraph_nodes, mafgraph_edges = mafgraph.get_graph_elements(jsonpangenome)
+    mafgraph_elements = []
+    mafgraph_elements.extend(mafgraph_nodes)
+    mafgraph_elements.extend(mafgraph_edges)
+    return mafgraph_elements
 
 @app.callback(
     Output(id_mafgraph_graph, 'style'),
