--- dash_app/components/texts.py
@@ -1,21 +1,0 @@
-algorithm_description = """
-## Pang
-
-Upload MAF to find consensuses in the multiple alignment.
-                                                
-[GitHub repository](https://github.com/meoke/pang)
-
-References:
-
-[Computational pan-genomics Marschall T. (2016). Computational pan-genomics: status, promises and challenges. 
-Briefiengs in Bioinformatics, pp. 1-18.](https://github.com/meoke/pang)
-
-[Lee C., Grasso C., Sharlow M.F. (2002). Multiple sequence alignment using partial order graphs.
-Bioinformatics, 18(3), pp. 452-464. ]()
-"""
-
-table_info_markdown = """
-Move the slider to select compatibility value respected in the below table.
- 
-This table includes only these consensuses which are the closest to the right side of the slider.
-"""
