--- dash_app/layout/about.py
@@ -1,4 +1,0 @@
-import dash_html_components as html
-
-def get_about_tab_content():
-    return html.Div("ABOUT")
