--- dash_app/layout/authors.py
@@ -1,5 +1,0 @@
-import dash_html_components as html
-
-
-def get_authors_tab_content():
-    return html.Div("Norbert Dojer, Paulina Dziadkiewicz")
