--- dash_app/layout/css_styles.py
@@ -1,12 +1,0 @@
-# class Colors:
-#     background='rgb(252,253,255)'
-#     light_background='rgb(179,207,226)'
-#     info='rgb(80,133,165)'
-#     dark_background='rgb(80,133,165)'
-#     accent='rgb(255,137,48)'
-#     error='rgb(150,3,26)'
-#     correct='rgb(105,120,101)'
-
-import dash_bootstrap_components as dbc
-
-
