+++ dash_app/old/tools.css
@@ -11,15 +11,7 @@
 /*    margin: 10px*/
 /*}*/
 
-.file_upload{
-    width:100%;
-    height:60px;
-    line-height: 60px;
-    border: 1px black dashed;
-    border-radius: 5px;
-    text-align: center;
-    /*margin: 10px;*/
-}
+
 
 /*#example_data_dropdown{*/
 /*    margin-bottom: 10px;*/
